가장 기본적인 요청 인증값 확인

+++ src/main/java/com/toyproject/toyproject/api/controller/PostController.java
@@ -20,9 +20,11 @@ public class PostController {
     private final PostService postService;
 
     @PostMapping("/posts")
-    public void post(@RequestBody @Valid PostCreate request) throws Exception {
-        request.validate();
-        postService.write(request);
+    public void post(@RequestBody @Valid PostCreate request, @RequestHeader String authorization) throws Exception {
+        if (authorization.equals("hodolman")) {
+            request.validate();
+            postService.write(request);
+        }
     }
 
     @GetMapping("/posts/{postId}")
@@ -36,12 +38,16 @@ public class PostController {
     }
 
     @PatchMapping("/posts/{postId}")
-    public void edit(@PathVariable Long postId, @RequestBody @Valid PostEdit postEdit) {
-        postService.edit(postId, postEdit);
+    public void edit(@PathVariable Long postId, @RequestBody @Valid PostEdit postEdit, @RequestHeader String authorization) {
+        if (authorization.equals("hodolman")) {
+            postService.edit(postId, postEdit);
+        }
     }
 
     @DeleteMapping("/posts/{postId}")
-    public void delete(@PathVariable Long postId) {
-        postService.delete(postId);
+    public void delete(@PathVariable Long postId, @RequestHeader String authorization) {
+        if (authorization.equals("hodolman")) {
+            postService.delete(postId);
+        }
     }
 }
